HitPause script
Simple script for applying hitpause to an action (eg getting hit or landing an attack). Will pause the game, changing timescale to 0. You should only pause for FRACTIONS OF A SECOND, it's usually not directly noticeable but ~feels good~.
(If we have too many scripts manipulating timescale, we may need a larger timescale manager script).